requirements.txt
dnspython==2.3.0
numpy==1.24.2
pandas==1.5.3
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sklearn==0.0.post1
PyYAML
scikit-learn
-e .